Rename to weekday2 to be renamed back
