+++ src/filters/weekday2/weekday2.constants.test.ts
@@ -1,4 +1,4 @@
-import WEEKDAY from './weekday.constants';
+import WEEKDAY from './weekday2.constants';
 
 describe('WEEKDAY', () => {
     describe('WITH_FUTURE_PAST_WORDS', () => {
